+++ src/render/renderer/pixels/renderer_pixels.h
@@ -6,7 +6,7 @@
 /*   By: arcanava <arcanava@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/05 17:45:38 by arcanava          #+#    #+#             */
-/*   Updated: 2025/01/14 15:33:45 by arcanava         ###   ########.fr       */
+/*   Updated: 2025/01/16 18:10:45 by arcanava         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@
 typedef t_iterators	t_pixel;
 
 t_pixel	*get_random_pixels(t_pixel *pixels, int parts_amount,
-			uint32_t *seed, t_size *w_size);
+			uint64_t *seed, t_size *w_size);
 
 void	fill_pixels(t_size img_size, t_pixel **pixels_, size_t *px_amount);
 
-void	shuffle_pixels(t_pixel *array, size_t n, uint32_t *seed);
+void	shuffle_pixels(t_pixel *array, size_t n, uint64_t *seed);
