+++ src/scene/figure/material/helpers/material_helpers.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   material_helpers.h                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cfidalgo <cfidalgo@student.42barcelona.    +#+  +:+       +#+        */
+/*   By: arcanava <arcanava@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/21 13:21:39 by cfidalgo          #+#    #+#             */
-/*   Updated: 2024/11/24 00:06:30 by cfidalgo         ###   ########.fr       */
+/*   Updated: 2025/01/16 18:10:45 by arcanava         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,13 +17,13 @@
 float	get_fresnel_effect(float index, float cos);
 
 void	randomize_ray_direction(t_vector *ideal_bounce_direction,
-			t_point *hit_point, uint32_t *seed, t_vector *res);
+			t_point *hit_point, uint64_t *seed, t_vector *res);
 
 int		diffuse_scatter(t_render *render, t_scatter_params *params,
-			t_color *direct_light, uint32_t *seed);
+			t_color *direct_light, uint64_t *seed);
 
 void	divert_ray_direction(t_vector *ray_dir, float roughness,
-			uint32_t *seed, t_vector *res);
+			uint64_t *seed, t_vector *res);
 
 int		metallic_scatter(t_render *render, t_scatter_params *params,
-			t_color *direct_light, uint32_t *seed);
+			t_color *direct_light, uint64_t *seed);
